// src/apis/generated/model/collection.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Unsplash API
 * Unsplash API specification for Alex Varelo's portfolio
 * OpenAPI spec version: 1.0.0
 */
import type { Tag } from './tag';
import type { Links } from './links';
import type { User } from './user';
import type { Photo } from './photo';

export interface Collection {
  id: string;
  title: string;
  description: string;
  published_at: string;
  last_collected_at: string;
  updated_at: string;
  featured: boolean;
  total_photos: number;
  private: boolean;
  share_key: string;
  tags: Tag[];
  links: Links;
  user: User;
  cover_photo: Photo;
  preview_photos: Photo[];
}